Split out `getDocumentHeadCappedUnreadCount`
